daemon: Length of message is a 32 bit unsigned quantity. (cherry picked from commit f65f653ab1541400afe2f796ab1b20cd721e769d)

+++ daemon/proto.c
@@ -312,7 +312,7 @@ reply (xdrproc_t xdrp, char *ret)
   char buf[GUESTFS_MESSAGE_MAX];
   char lenbuf[4];
   struct guestfs_message_header hdr;
-  unsigned len;
+  uint32_t len;
 
   xdrmem_create (&xdr, buf, sizeof buf, XDR_ENCODE);
 
@@ -352,7 +352,7 @@ reply (xdrproc_t xdrp, char *ret)
     fprintf (stderr, "guestfsd: xwrite failed\n");
     exit (EXIT_FAILURE);
   }
-  if (xwrite (sock, buf, len) == -1) {
+  if (xwrite (sock, buf, (size_t) len) == -1) {
     fprintf (stderr, "guestfsd: xwrite failed\n");
     exit (EXIT_FAILURE);
   }
